refactor : packaging memberService
